split object files
